Use per-client auth keys, instead of a global auth key.
In the past prop224 used to embed the client authorization key in the
subcredential. The problem here is that if we wanted to revoke a client,
we would have to change the whole subcredential of the service, which
means that we would have to announce it to all clients.

This patch makes it so that every client has their own authorization
key. We use the old rend-spec.txt trick of encrypting the descriptor key
to each individual client authorization key here as well. This makes
revocation practical again.

Thanks to special for noticing this issue.