prop224: Use per-client desc auth keys, instead of global desc auth key.
In the past prop224 used to embed the client authorization key in the
subcredential. The problem here is that if we wanted to revoke a client,
we would have to change the whole subcredential of the service, which
means that we would have to announce it to all clients.

This patch makes it so that every client has their own authorization
key. We use the old rend-spec.txt trick of encrypting the descriptor key
to each individual client authorization key here as well. This makes
revocation practical again.

To achieve this we change the descriptor format to a double-layer
encryption where the first layer protects against entities who don't
know the public key of the HS, and the second layer protects against
unauthorized clients.

Thanks to special for noticing this issue.