fixup! Appendix to rend-spec.txt about how to generate revision counters
We cannot use "seconds since time period started" for rev counter, because for
some descriptors onion services will use the next time period, and that would
cause the OPE plaintext to be negative. See SERVICEUPLOAD in rend-spec-v3.txt.

Hence we use the "seconds since SRV got published" which is guaranteed to be
positive (since we need an SRV to make a descriptor).