Shrink cell nonce to 16 bytes after tevador's analysis.
I suggest to use a 16-byte nonce value, which is more than sufficient
given the
target security level and has the benefit of reducing the replay cache
size to
half.

Since we expect the seed to be valid for around 3 hours (as proposed),
then even
if the service receives 1 million proofs per second and each proof has
an effort
of 1 million, then the number of submitted nonces from clients will
only reach
about 10^10. With a 108-bit solution space (subtracting 20 bits as the
search
space per client), the probability that two clients accidentally submit
the same
nonce is roughly 10^-13 (see [REF_BIRTHDAY]).